index.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>Masked Input tests</title>
    <meta name="timeout" content="long" />
    <meta name="color-scheme" content="dark light" />
    <!-- polyfill for safari -->
    <script src="/assets/polyfill/custom-elements.js"></script>
    <!-- Web Platform Tests (WPT) testing suite -->
    <script src="/assets/wpt/testharness.js"></script>
    <script src="/assets/wpt/testharnessreport.js"></script>
    <script src="/assets/masked-input.js"></script>
  </head>
  <body>
    <input type="password" is="masked-input" mask-pattern="___ ___ ___" />
    <input type="number" is="masked-input" mask-pattern="___ ___ ___" />
    <script>
      function createMaskedInput({ props = {}, attrs = {} }) {
        const maskedInput = document.createElement("input", {
          is: "masked-input",
        });

        for (const [attrName, attrValue] of Object.entries(attrs)) {
          maskedInput.setAttribute(attrName, attrValue ?? "");
        }
        for (const [propName, propValue] of Object.entries(props)) {
          if (propValue != null) {
            maskedInput[propName] = propValue;
          }
        }

        document.body.appendChild(maskedInput);

        return maskedInput;
      }

      /**
       * 1. masked -> unmasked for delete events, unmasked -> masked for insert events
       * 2. no change in mask visibility
       * 3. beginning, middle, and end positions
       */
      const testCases = [
        {
          inputTypes: [
            "deleteContent",
            "deleteContentBackward",
            "deleteByDrag",
            "deleteByCut",
          ],
          baseInput: {
            maskPattern: "-_-_-_-",
            value: "1234",
            selectionStart: 0,
            selectionEnd: 0,
            attributes: {}, // 'show-empty-mask' etc.
            data: null,
          },
          variants: [
            {
              description: "noop: delete backward when caret is at beginning",
              input: {},
              output: {
                value: "1234",
                maskedValue: "1234",
                selectionStart: 0,
                selectionEnd: 0,
              },
            },
            {
              description: "UNMASKED -> MASKED: delete backward from end",
              input: {
                selectionStart: 4,
                selectionEnd: 4,
              },
              output: {
                value: "123",
                maskedValue: "-1-2-3-",
                selectionStart: 6,
                selectionEnd: 6,
              },
            },
            {
              description: "UNMASKED -> UNMASKED: delete backward from middle",
              input: {
                value: "12345",
                selectionStart: 4,
                selectionEnd: 4,
              },
              output: {
                value: "1235",
                maskedValue: "1235",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
            {
              description: "MASKED -> MASKED: delete backward from end",
              input: {
                value: "123",
                selectionStart: 4,
                selectionEnd: 4,
              },
              output: {
                value: "13",
                maskedValue: "-1-3- -",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
            {
              description:
                "UNMASKED -> MASKED: delete backward from end, showing replacement characters",
              input: {
                value: "123",
                selectionStart: 4,
                selectionEnd: 4,
                attributes: {
                  "show-replacement-characters": true,
                },
              },
              output: {
                value: "13",
                maskedValue: "-1-3-_-",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
            {
              description:
                "MASKED -> UNMASKED: delete backward from end, going from single character to empty value",
              input: {
                value: "1",
                selectionStart: 2,
                selectionEnd: 2,
              },
              output: {
                value: "",
                maskedValue: "",
                selectionStart: 0,
                selectionEnd: 0,
              },
            },
            {
              description:
                "delete selection ([-1-2-3-]), going from masked to unmasked empty value",
              input: {
                value: "123",
                selectionStart: 0,
                selectionEnd: 6,
              },
              output: {
                value: "",
                maskedValue: "",
                selectionStart: 0,
                selectionEnd: 0,
              },
            },
            {
              description:
                "delete selection (-1[-2-3-]), going from masked to masked",
              input: {
                value: "123",
                selectionStart: 2,
                selectionEnd: 6,
              },
              output: {
                value: "1",
                maskedValue: "-1- - -",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
            {
              description:
                "delete selection (-1-[2-3-]), going from masked to masked",
              input: {
                value: "123",
                selectionStart: 3,
                selectionEnd: 6,
              },
              output: {
                value: "1",
                maskedValue: "-1- - -",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
            {
              description:
                "delete backward from end, going from masked single character to masked empty value",
              input: {
                value: "1",
                selectionStart: 2,
                selectionEnd: 2,
                attributes: {
                  "show-empty-mask": "true",
                },
              },
              output: {
                value: "",
                maskedValue: "- - - -",
                selectionStart: 1,
                selectionEnd: 1,
              },
            },
          ],
        },
        {
          inputTypes: ["deleteContentForward"],
          baseInput: {
            maskPattern: "-_-_-_-",
            value: "1234",
            selectionStart: 0,
            selectionEnd: 0,
          },
          variants: [
            {
              description: "unmasked to masked: delete single character",
              input: {},
              output: {
                value: "234",
                maskedValue: "-2-3-4-",
                selectionStart: 1,
                selectionEnd: 1,
              },
            },
            {
              description: "noop: caret at end of input",
              input: {
                selectionStart: 4,
                selectionEnd: 4,
              },
              output: {
                value: "1234",
                maskedValue: "1234",
                selectionStart: 4,
                selectionEnd: 4,
              },
            },
            {
              description: "delete single character, caret does not move",
              input: {
                value: "12345",
                selectionStart: 4,
                selectionEnd: 4,
              },
              output: {
                value: "1234",
                maskedValue: "1234",
                selectionStart: 4,
                selectionEnd: 4,
              },
            },
            {
              description:
                "delete single character, following characters move backward",
              input: {
                value: "123",
                selectionStart: 3,
                selectionEnd: 3,
              },
              output: {
                value: "13",
                maskedValue: "-1-3- -",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
            {
              description:
                "delete single character, replaced with replacement character",
              input: {
                value: "123",
                selectionStart: 3,
                selectionEnd: 3,
                attributes: {
                  "show-replacement-characters": true,
                },
              },
              output: {
                value: "13",
                maskedValue: "-1-3-_-",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
            {
              description: "delete entire input (single character)",
              input: {
                value: "1",
                selectionStart: 1,
                selectionEnd: 1,
              },
              output: {
                value: "",
                maskedValue: "",
                selectionStart: 0,
                selectionEnd: 0,
              },
            },
            {
              description: "delete selection",
              input: {
                value: "123",
                selectionStart: 0,
                selectionEnd: 6,
              },
              output: {
                value: "",
                maskedValue: "",
                selectionStart: 0,
                selectionEnd: 0,
              },
            },
            {
              description: "delete partial selection",
              input: {
                value: "123",
                selectionStart: 2,
                selectionEnd: 6,
              },
              output: {
                value: "1",
                maskedValue: "-1- - -",
                selectionStart: 2,
                selectionEnd: 2,
              },
            },
            {
              description:
                "delete partial selection, caret does not move from start position",
              input: {
                value: "123",
                selectionStart: 3,
                selectionEnd: 6,
              },
              output: {
                value: "1",
                maskedValue: "-1- - -",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
            {
              description: "delete entire input (single character), show mask",
              input: {
                value: "1",
                selectionStart: 1,
                selectionEnd: 1,
                attributes: {
                  "show-empty-mask": "true",
                },
              },
              output: {
                value: "",
                maskedValue: "- - - -",
                selectionStart: 1,
                selectionEnd: 1,
              },
            },
          ],
        },
        {
          inputTypes: ["deleteWordBackward"],
          baseInput: {
            maskPattern: "-_-_-_-_-_-",
            value: "12 34",
            selectionStart: 10,
            selectionEnd: 10,
          },
          variants: [
            {
              description: "noop: delete from start",
              input: {
                selectionStart: 1,
                selectionEnd: 1,
              },
              output: {
                value: "12 34",
                maskedValue: "-1-2- -3-4-",
                selectionStart: 1,
                selectionEnd: 1,
              },
            },
            {
              description: "delete from middle",
              input: {
                selectionStart: 6,
                selectionEnd: 6,
              },
              output: {
                value: "34",
                maskedValue: "-3-4- - - -",
                selectionStart: 1,
                selectionEnd: 1,
              },
            },
            {
              description: "delete from end",
              input: {},
              output: {
                value: "12 ",
                maskedValue: "-1-2- - - -",
                selectionStart: 7,
                selectionEnd: 7,
              },
            },
            {
              description: "delete from middle of word",
              input: {
                selectionStart: 8,
                selectionEnd: 8,
              },
              output: {
                value: "12 4",
                maskedValue: "-1-2- -4- -",
                selectionStart: 7,
                selectionEnd: 7,
              },
            },
            {
              description: "delete entire input (one word with trailing space)",
              input: {
                value: "12 ",
                selectionStart: 6,
                selectionEnd: 6,
              },
              output: {
                value: "",
                maskedValue: "",
                selectionStart: 0,
                selectionEnd: 0,
              },
            },
            {
              description: "delete only word, leaving leading space",
              input: {
                value: " 234",
                maskPattern: "-_-",
                selectionStart: 4,
                selectionEnd: 4,
              },
              output: {
                value: " ",
                maskedValue: "- -",
                selectionStart: 2,
                selectionEnd: 2,
              },
            },
            {
              description: "delete selection",
              input: {
                selectionStart: 3,
                selectionEnd: 8,
              },
              output: {
                value: "14",
                maskedValue: "-1-4- - - -",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
          ],
        },
        {
          inputTypes: ["deleteWordForward"],
          baseInput: {
            maskPattern: "-_-_-_-_-_-",
            value: "12 34",
            selectionStart: 4,
            selectionEnd: 4,
          },
          variants: [
            {
              description: "noop: caret at end of input",
              input: {
                selectionStart: 10,
                selectionEnd: 10,
              },
              output: {
                value: "12 34",
                maskedValue: "-1-2- -3-4-",
                selectionStart: 10,
                selectionEnd: 10,
              },
            },
            {
              description: "delete first word and not the space after it",
              input: {
                selectionStart: 1,
                selectionEnd: 1,
              },
              output: {
                value: " 34",
                maskedValue: "- -3-4- - -",
                selectionStart: 1,
                selectionEnd: 1,
              },
            },
            {
              description: "delete second word and space before",
              input: {},
              output: {
                value: "12",
                maskedValue: "-1-2- - - -",
                selectionStart: 4,
                selectionEnd: 4,
              },
            },
            {
              description: "delete partial word",
              input: {
                selectionStart: 8,
                selectionEnd: 8,
              },
              output: {
                value: "12 3",
                maskedValue: "-1-2- -3- -",
                selectionStart: 8,
                selectionEnd: 8,
              },
            },
            {
              description:
                "delete the entire value (a single word following a space)",
              input: {
                value: " 12",
                selectionStart: 1,
                selectionEnd: 1,
              },
              output: {
                value: "",
                maskedValue: "",
                selectionStart: 0,
                selectionEnd: 0,
              },
            },
            {
              description: "umasked to masked",
              input: {
                value: "234 ",
                maskPattern: "-_-",
                selectionStart: 0,
                selectionEnd: 0,
              },
              output: {
                value: " ",
                maskedValue: "- -",
                selectionStart: 1,
                selectionEnd: 1,
              },
            },
            {
              description: "delete selection (ignores word boundaries)",
              input: {
                selectionStart: 3,
                selectionEnd: 8,
              },
              output: {
                value: "14",
                maskedValue: "-1-4- - - -",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
          ],
        },
        {
          inputTypes: ["deleteWordForward"],
          baseInput: {
            maskPattern: "-_-_-_-_-_-",
            value: "12 34",
            selectionStart: 1,
            selectionEnd: 1,
          },
          variants: [
            {
              description: "noop: delete from end",
              input: {
                selectionStart: 10,
                selectionEnd: 10,
              },
              output: {
                value: "12 34",
                maskedValue: "-1-2- -3-4-",
                selectionStart: 10,
                selectionEnd: 10,
              },
            },
            {
              description: "delete from middle, before space",
              input: {
                selectionStart: 4,
                selectionEnd: 4,
              },
              output: {
                value: "12",
                maskedValue: "-1-2- - - -",
                selectionStart: 4,
                selectionEnd: 4,
              },
            },
            {
              description: "delete from start",
              input: {},
              output: {
                value: " 34",
                maskedValue: "- -3-4- - -",
                selectionStart: 1,
                selectionEnd: 1,
              },
            },
            {
              description: "delete from middle of word",
              input: {
                selectionStart: 8,
                selectionEnd: 8,
              },
              output: {
                value: "12 3",
                maskedValue: "-1-2- -3- -",
                selectionStart: 8,
                selectionEnd: 8,
              },
            },
            {
              description: "delete entire input (one word with leading space)",
              input: {
                value: " 12",
                selectionStart: 1,
                selectionEnd: 1,
              },
              output: {
                value: "",
                maskedValue: "",
                selectionStart: 0,
                selectionEnd: 0,
              },
            },
            {
              description: "delete only word, leaving trailing space",
              input: {
                value: "123 ",
                maskPattern: "-_-",
                selectionStart: 0,
                selectionEnd: 0,
              },
              output: {
                value: " ",
                maskedValue: "- -",
                selectionStart: 1,
                selectionEnd: 1,
              },
            },
            {
              description: "delete selection",
              input: {
                selectionStart: 3,
                selectionEnd: 8,
              },
              output: {
                value: "14",
                maskedValue: "-1-4- - - -",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
          ],
        },
        {
          inputTypes: ["deleteSoftLineForward", "deleteHardLineForward"],
          baseInput: {
            maskPattern: "-_-_-_-_-",
            value: "1234",
            selectionStart: 1,
            selectionEnd: 1,
          },
          variants: [
            {
              description: "delete from start",
              input: {},
              output: {
                value: "",
                maskedValue: "",
                selectionStart: 0,
                selectionEnd: 0,
              },
            },
            {
              description: "noop: delete from end",
              input: {
                selectionStart: 8,
                selectionEnd: 8,
              },
              output: {
                value: "1234",
                maskedValue: "-1-2-3-4-",
                selectionStart: 8,
                selectionEnd: 8,
              },
            },
            {
              description: "delete from middle (at end of 1st char)",
              input: {
                selectionStart: 2,
                selectionEnd: 2,
              },
              output: {
                value: "1",
                maskedValue: "-1- - - -",
                selectionStart: 2,
                selectionEnd: 2,
              },
            },
            {
              description: "delete from middle (at start of 2nd char)",
              input: {
                selectionStart: 3,
                selectionEnd: 3,
              },
              output: {
                value: "1",
                maskedValue: "-1- - - -",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
            {
              description: "unmasked to masked: delete from middle",
              input: {
                value: "12",
                maskPattern: "-_-",
                selectionStart: 1,
                selectionEnd: 1,
              },
              output: {
                value: "1",
                maskedValue: "-1-",
                selectionStart: 2,
                selectionEnd: 2,
              },
            },
          ],
        },
        {
          inputTypes: ["deleteSoftLineBackward", "deleteHardLineBackward"],
          baseInput: {
            maskPattern: "-_-_-_-_-",
            value: "1234",
            selectionStart: 1,
            selectionEnd: 1,
          },
          variants: [
            {
              description: "delete from end",
              input: {
                selectionStart: 8,
                selectionEnd: 8,
              },
              output: {
                value: "",
                maskedValue: "",
                selectionStart: 0,
                selectionEnd: 0,
              },
            },
            {
              description: "noop: caret at start",
              input: {},
              output: {
                value: "1234",
                maskedValue: "-1-2-3-4-",
                selectionStart: 1,
                selectionEnd: 1,
              },
            },
            {
              description: "delete from middle (at end of 1st char)",
              input: {
                selectionStart: 2,
                selectionEnd: 2,
              },
              output: {
                value: "234",
                maskedValue: "-2-3-4- -",
                selectionStart: 1,
                selectionEnd: 1,
              },
            },
            {
              description: "delete from middle (at start of 2nd char)",
              input: {
                selectionStart: 3,
                selectionEnd: 3,
              },
              output: {
                value: "234",
                maskedValue: "-2-3-4- -",
                selectionStart: 1,
                selectionEnd: 1,
              },
            },
            {
              description: "unmasked to masked: delete from middle",
              input: {
                value: "12",
                maskPattern: "-_-",
                selectionStart: 1,
                selectionEnd: 1,
              },
              output: {
                value: "2",
                maskedValue: "-2-",
                selectionStart: 1,
                selectionEnd: 1,
              },
            },
          ],
        },
        {
          inputTypes: ["deleteEntireSoftLine"],
          baseInput: {
            maskPattern: "-_-_-_-_-",
            value: "1234",
            selectionStart: 1,
            selectionEnd: 1,
          },
          variants: [
            {
              description: "delete from end",
              input: {
                selectionStart: 8,
                selectionEnd: 8,
              },
              output: {
                value: "",
                maskedValue: "",
                selectionStart: 0,
                selectionEnd: 0,
              },
            },
            {
              description: "delete from start",
              input: {},
              output: {
                value: "",
                maskedValue: "",
                selectionStart: 0,
                selectionEnd: 0,
              },
            },
            {
              description: "delete from middle",
              input: {
                selectionStart: 2,
                selectionEnd: 2,
              },
              output: {
                value: "",
                maskedValue: "",
                selectionStart: 0,
                selectionEnd: 0,
              },
            },
          ],
        },
        {
          inputTypes: [
            "insertReplacementText",
            "insertFromYank",
            "insertFromDrop",
            "insertFromPaste",
            "insertFromPasteAsQuotation",
            "insertCompositionText",
            "insertText",
          ],
          baseInput: {
            maskPattern: "-_-_-_-",
            value: "12",
            data: "3",
            selectionStart: 4,
            selectionEnd: 4,
          },
          variants: [
            {
              description: "insert at start",
              input: {
                selectionStart: 1,
                selectionEnd: 1,
              },
              output: {
                value: "312",
                maskedValue: "-3-1-2-",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
            {
              description: "insert at end",
              input: {},
              output: {
                value: "123",
                maskedValue: "-1-2-3-",
                selectionStart: 6,
                selectionEnd: 6,
              },
            },
            {
              description: "insert in middle (at end of 1st char)",
              input: {
                selectionStart: 2,
                selectionEnd: 2,
              },
              output: {
                value: "132",
                maskedValue: "-1-3-2-",
                selectionStart: 4,
                selectionEnd: 4,
              },
            },
            {
              description: "insert in middle (at start of 2nd char)",
              input: {
                selectionStart: 3,
                selectionEnd: 3,
              },
              output: {
                value: "132",
                maskedValue: "-1-3-2-",
                selectionStart: 5,
                selectionEnd: 5,
              },
            },
            {
              description: "unmasked to masked: insert into empty input",
              input: {
                value: "",
                maskPattern: "-_-_-",
                selectionStart: 0,
                selectionEnd: 0,
              },
              output: {
                value: "3",
                maskedValue: "-3- -",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
            {
              description:
                "masked to unmasked: insert into full input (overflow mask)",
              input: {
                value: "1",
                maskPattern: "-_-",
                selectionStart: 2,
                selectionEnd: 2,
              },
              output: {
                value: "13",
                maskedValue: "13",
                selectionStart: 2,
                selectionEnd: 2,
              },
            },
            {
              description:
                "masked to unmasked: insert into full input (show overflow mask)",
              input: {
                value: "1",
                maskPattern: "-_-",
                selectionStart: 2,
                selectionEnd: 2,
                attributes: {
                  "show-overflowed-mask": true,
                },
              },
              output: {
                value: "13",
                maskedValue: "-1-3",
                selectionStart: 4,
                selectionEnd: 4,
              },
            },
            {
              description: "unmasked to masked: insert grapheme cluster",
              input: {
                value: "",
                data: "🇺🇸",
                maskPattern: "-_-",
                selectionStart: 0,
                selectionEnd: 0,
              },
              output: {
                value: "🇺🇸",
                maskedValue: "-🇺🇸-",
                selectionStart: 5,
                selectionEnd: 5,
              },
            },
            {
              description: "unmasked to masked: insert over selection",
              input: {
                value: "1234",
                data: "x",
                maskPattern: "-_-_-_-",
                selectionStart: 1,
                selectionEnd: 3,
              },
              output: {
                value: "1x4",
                maskedValue: "-1-x-4-",
                selectionStart: 5,
                selectionEnd: 5,
              },
            },
          ],
        },
        {
          inputTypes: ["insertTranspose"],
          baseInput: {
            maskPattern: "-_-_-",
            value: "12",
          },
          variants: [
            {
              description: "noop: caret at start",
              input: {
                selectionStart: 1,
                selectionEnd: 1,
              },
              output: {
                value: "12",
                maskedValue: "-1-2-",
                selectionStart: 1,
                selectionEnd: 1,
              },
            },
            {
              description:
                "transpose between two chars (at end of 1st char) separated by a mask char",
              input: {
                selectionStart: 2,
                selectionEnd: 2,
              },
              output: {
                value: "21",
                maskedValue: "-2-1-",
                selectionStart: 4,
                selectionEnd: 4,
              },
            },
            {
              description:
                "transpose between two chars (at start of 2nd char) separated by a mask char",
              input: {
                selectionStart: 3,
                selectionEnd: 3,
              },
              output: {
                value: "21",
                maskedValue: "-2-1-",
                selectionStart: 4,
                selectionEnd: 4,
              },
            },
            {
              description:
                "transpose between two chars (at end of 2nd char) separated by a mask char",
              input: {
                selectionStart: 4,
                selectionEnd: 4,
              },
              output: {
                value: "21",
                maskedValue: "-2-1-",
                selectionStart: 4,
                selectionEnd: 4,
              },
            },
            {
              description:
                "transpose between two chars not separated by a mask char, caret in middle",
              input: {
                selectionStart: 2,
                selectionEnd: 2,
                maskPattern: "-__-",
              },
              output: {
                value: "21",
                maskedValue: "-21-",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
            {
              description:
                "transpose between two chars not separated by a mask char, caret at end",
              input: {
                selectionStart: 3,
                selectionEnd: 3,
                maskPattern: "-__-",
              },
              output: {
                value: "21",
                maskedValue: "-21-",
                selectionStart: 3,
                selectionEnd: 3,
              },
            },
          ],
        },
      ];

      customElements.whenDefined("masked-input").then(() => {
        const types = [
          "text",
          "url",
          "search",
          "tel",
          "email",
          "password",
          "number",
        ];
        const attributes = [
          "mask-pattern",
          "mask-replacement-character",
          "show-empty-mask",
          "show-overflowed-mask",
          "show-replacement-characters",
          "type",
          "value",
          "maxlength",
          "minlength",
          "pattern",
          "inputmode",
          "required",
          "step",
          "min",
          "max",
        ];
        const props = [];
        // configurations
        test(() => {
          const m1 = createMaskedInput({
            props: {
              maskPattern: "<<_>>",
              value: "",
            },
            attrs: {
              "show-empty-mask": "",
            },
          });
          assert_equals(m1.maskedValue, "<< >>", "with props");
          m1.remove();

          const m2 = createMaskedInput({
            attrs: {
              "show-empty-mask": "",
              "mask-pattern": "<<_>>",
              value: "",
            },
          });
          assert_equals(m2.maskedValue, "<< >>", "with attributes");
          m2.remove();

          const m3 = createMaskedInput({
            attrs: {
              "show-empty-mask": "",
              "show-replacement-characters": "",
              "mask-pattern": "<<_>>",
              value: "",
            },
          });
          assert_equals(
            m3.maskedValue,
            "<<_>>",
            "showing replacement characters"
          );
          m3.remove();

          const m4 = createMaskedInput({
            attrs: {
              "show-empty-mask": "",
              "show-replacement-characters": "",
              "mask-replacement-character": "X",
              "mask-pattern": "<<X>>",
              value: "",
            },
          });
          assert_equals(
            m4.maskedValue,
            "<<X>>",
            "showing custom (attr) replacement characters"
          );
          m4.remove();

          const m5 = createMaskedInput({
            props: {
              maskReplacementCharacter: "X",
            },
            attrs: {
              "show-empty-mask": "",
              "show-replacement-characters": "",
              "mask-pattern": "<<X>>",
              value: "",
            },
          });
          assert_equals(
            m5.maskedValue,
            "<<X>>",
            "showing custom (prop) replacement characters"
          );
          m5.remove();

          const m6 = createMaskedInput({
            props: {
              maskReplacementCharacter: "X",
            },
            attrs: {
              "mask-pattern": "<<X>>",
              value: "",
            },
          });
          assert_equals(m6.maskedValue, "", "defaults to empty display value");
          m6.remove();
        }, "empty masked input in various configurations");

        test(() => {
          assert_throws_dom(
            "NotSupportedError",
            () => {
              createMaskedInput({ props: { type: "range" } });
            },
            "prop"
          );
          assert_throws_dom(
            "NotSupportedError",
            () => {
              createMaskedInput({ attrs: { type: "range" } });
            },
            "attr"
          );
        }, "invalid configurations");

        test(() => {
          const m1 = createMaskedInput({
            props: {
              maskPattern: "<<_>>",
              value: "1",
            },
          });
          assert_equals(m1.maskedValue, "<<1>>", "with props");
          m1.remove();

          const m2 = createMaskedInput({
            attrs: {
              "mask-pattern": "<<_>>",
              value: "2",
            },
          });
          assert_equals(m2.maskedValue, "<<2>>", "with attributes");
          m2.remove();

          const m3 = createMaskedInput({
            attrs: {
              "show-replacement-characters": "",
              "mask-pattern": "<_> {_}",
              value: "3",
            },
          });
          assert_equals(
            m3.maskedValue,
            "<3> {_}",
            "showing replacement characters"
          );
          m3.remove();

          const m4 = createMaskedInput({
            attrs: {
              "show-replacement-characters": "",
              "mask-replacement-character": "X",
              "mask-pattern": "<X> {X}",
              value: "4",
            },
          });
          assert_equals(
            m4.maskedValue,
            "<4> {X}",
            "showing custom (attr) replacement characters"
          );
          m4.remove();

          const m5 = createMaskedInput({
            props: {
              maskReplacementCharacter: "X",
            },
            attrs: {
              "show-replacement-characters": "",
              "mask-pattern": "<X> {X}",
              value: "5",
            },
          });
          assert_equals(
            m5.maskedValue,
            "<5> {X}",
            "showing custom (prop) replacement characters"
          );
          m5.remove();
        }, "filled masked input in various configurations");

        test(() => {
          const m1 = createMaskedInput({
            props: {
              maskPattern: "<<_>>",
              value: "123",
            },
          });
          assert_equals(m1.maskedValue, "123", "with props");
          m1.remove();

          const m2 = createMaskedInput({
            attrs: {
              "mask-pattern": "<<_>>",
              value: "123",
            },
          });
          assert_equals(m2.maskedValue, "123", "with attributes");
          m2.remove();

          const m3 = createMaskedInput({
            attrs: {
              "show-overflowed-mask": "",
              "mask-pattern": "<_> {_}",
              value: "123",
            },
          });
          assert_equals(m3.maskedValue, "<1> {2}3", "showing overflowed mask");
          m3.remove();
        }, "overflowed masked input in various configurations");

        test(() => {
          Object.entries({
            text: "text",
            url: "url",
            search: "search",
            tel: "tel",
            email: "email",
            password: "text",
            number: "decimal",
          }).forEach(([type, inputmode]) => {
            const m1 = createMaskedInput({ props: { type } });
            assert_equals(m1.inputMode, inputmode, `[type: ${type}]`);
            m1.remove();

            const m2 = createMaskedInput({ attrs: { type } });
            assert_equals(m2.inputMode, inputmode, `[type: ${type}]`);
            m2.remove();
          });
        }, "inputmode attribute by type");

        // input tests
        // for (const { inputTypes, baseInput, variants } of testCases) {
        //   for (const { input, output, description = "" } of variants) {
        //     for (const inputType of inputTypes) {
        //       for (const selectionDirection of [
        //         "forward",
        //         "backward",
        //         "none",
        //       ]) {
        //         for (const type of ["text", "url", "search", "tel", "email"]) {
        //           promise_test(async () => {
        //             const {
        //               maskPattern,
        //               maskReplacementCharacter,
        //               value,
        //               attributes = {},
        //               selectionStart,
        //               selectionEnd,
        //               data = null,
        //             } = { ...baseInput, ...input };
        //             const maskedInput = createMaskedInput({
        //               props: {
        //                 type,
        //                 maskPattern,
        //                 maskReplacementCharacter,
        //                 value,
        //               },
        //               attrs: attributes,
        //             });
        //             maskedInput.setSelectionRange(
        //               selectionStart,
        //               selectionEnd,
        //               selectionDirection
        //             );
        //             maskedInput.dispatchEvent(
        //               new InputEvent("beforeinput", { inputType, data })
        //             );

        //             await new Promise((resolve) => {
        //               // wait for event handlers to run, then assert output
        //               setTimeout(() => {
        //                 assert_equals(
        //                   output.maskedValue,
        //                   maskedInput.maskedValue
        //                 );
        //                 assert_equals(output.value, maskedInput.value);
        //                 assert_equals(
        //                   output.selectionStart,
        //                   maskedInput.selectionStart
        //                 );
        //                 assert_equals(
        //                   output.selectionEnd,
        //                   maskedInput.selectionEnd
        //                 );

        //                 resolve();
        //               }, 0);
        //             });

        //             maskedInput.remove();
        //           }, type + ": " + inputType + ": selectionDirection: " + selectionDirection + ": " + description);
        //         }
        //       }
        //     }
        //   }
        // }
      });
    </script>
  </body>
</html>
